Update patch set 2
Patch Set 2:

The dell class does not require the ipmi class, and there are cases when you would not want the ipmi class to be automatically included (as it changes IP addresses, etc) and where you would want the dell repo and dell monitoring. The dell service requires the ipmi modules to be loaded, but nothing else from the ipmi class.

Patch-set: 2
Label: Verified=0