Update patch set 1
Patch Set 1:

This fact is throwing an error on RHEL because lscpi full path is /usr/sbin/lspci.

We do a quick hack for now. This fact should be remove in time because the only nodes that needs this fact (is machine a HP?) is nectar swift, and it is not efficient to run this fact on all nectar hosts to determine whether they are HP or not.

Patch-set: 1