Update patch set 1
Patch Set 1:

Calling lscpi without full path works. Also, we found out that %x{} has been 'replaced by Facter::Util::Resolution.exec'

https://puppetlabs.com/blog/facter-part-1-facter-101

Not sure how far we want to go to fix this. As per previous comment, this is a quick fix for now and we should move to do away with this fact.

Patch-set: 1