Puppet3to4: Fix up check for $ipmi_dns_lookup fact
Fact: In Puppet3, empty strings are false, but in Puppet4 they are true.
undef is always false. [1]

In 'lib/facter/ipmi.rb' the fact $ipmi_dns_lookup will be undef if the
puppet client can't find the ip address for it's ipmi hostname.

This check used to work in Puppet3, because when $ipmi_dns_lookup is
undef, it is (I think) comparing (undef != false), but now in Puppet4 it
is comparing (undef != true)

[1] https://docs.puppet.com/puppet/3.8/lang_datatypes.html#automatic-conversion-to-boolean

Change-Id: Ic06dca03d42ac8ceb9d7f04a9c42da71d03608f5