Allow specifying a distribution for hwraid repo
Although the packages largely work on Focal the hwraid project has not made a
focal repository available.

This is tracked in Github and there is mention of Python 3 being a factor,
however thats somewhat tangential for us.
https://github.com/eLvErDe/hwraid/issues/116

In commit 7b1d02da4675c6627175a82984b1fb082c22820b I changed hwraid to use
$::lsbdistcodename because it was previously hard coded to precise. I'm now
moving that $::lsbdistcodename usage a little so it can be overridden in hiera.

hwraid is moved to a new file to pass linting.

Change-Id: I7cdde5f3569a14f47a80533286aef4eb90e134ff