Explicitly set ipmi password as 16-byte
The ipmi_user_setpw exec resource checks whether the ipmi user has
a password stored as 16-byte. If not, it runs the set password
command.
On some hosts, the set password command defaults to storing the
password as 20-byte. On hosts that have ended up in this state
the ipmi_user_setpw exec resource is performed on every puppet run.
This current change explicitly sets the new password as 16-byte so
that subsequent checks succeed and the password does not get reset
repeatedly.

Change-Id: I0db9f01b4c021aeabd775a46728e4ba4eb604282