Mark the ipmi exec command to sensitive
Without this, octocatalog detects that the 'command' resource param
has changed and outputs a diff containing the password.

If we only set password as Sensitive, it does not help because after
unwrapping password in the 'command' param, the 'command' param is
still a plaintext string. otocatalog-diff only masks resource params
that are sensitive[1].

So we make the entire exec 'command' param sensitive, which hides it
in logs and octacatalog-diff output.

[1] Change Iba0748660560886eaeb98bc0aaeb9594b14f8e4a

Change-Id: I9c08789f05e89e788dcd8595adede1a5636d68e5