Make lowercase 'admin' user name work
Hitachi IPMI do not allow uppercase user names.

Downcasing user admin to match both admin and ADMIN. This needs the
following hiera set

 physical::ipmi::user: 'admin'

NOTE: puppet string comparison is insensitive [1], so we do not really
need to downcase it, just doing this for safety.

This class probably should be refactored to be more flexible, allowing
users to choose user number.

[1] https://www.puppet.com/docs/puppet/7/lang_expressions.html#lang_exp_comparison_operators

Change-Id: Iff65d42ae5e56e66b3571126d84f47355b99f008